requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
PyJWT==2.7.0
pytest==7.3.1
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
